requirements.txt
certifi==2019.6.16
chardet==4.0.0
dnspython==1.16.0
eventlet==0.31.0
greenlet==0.4.15
idna==2.8
monotonic==1.5
python-engineio==3.9.3
python-socketio==4.3.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
urllib3==1.26.8
minishogilib==0.6.12
